+++ ui/apps/platform/src/Containers/VulnMgmt/Entity/Component/VulnMgmtEntityComponent.js
@@ -32,6 +32,7 @@ const VulnMgmtComponent = ({
                 id
                 name
                 version
+                fixedIn
                 location(query: $scopeQuery)
                 priority
                 vulnCount(query: $query)