+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
 
 - The product no longer requires a license to run. Several license-related functionalities and flags
   have been removed from the product and related tooling, as well as from the Helm charts.
+- Components now have `Fixed By` field that indicates the version that will fixes all the fixable vulnerabilities in the component. 
+  - Note: 
+    - It is supported only when StackRox Scanner is used.
+    - It is not namespaced to distro.
 
 ## [57.0]